Change specification of input files
Replace 'counts.yaml' with 'data.yaml', which provides relative paths to
input data files and the corresponding numbers of entries. With this
change, don't need to impose a specific directory structure. The order
of files is fixed and does not depend on implementation details of how a
dictionary is loaded from YAML.