Include numeric properties of charged constituents
Have to change the padding value from NaN. Overwise the gradient with
respect to parameters of the dense subnetwork applied to charged
constituents becomes NaN as well. This is not cured by using
multiply_no_nan in MaskedSum because the parameters of the dense
subnetwork are multipied by NaNs before entrering MaskedSum.

Batch normalization layers will include the masked-out inputs when
computing their running statistics, which is not correct.